Updated dx_operations_vdb
dx_operations_vdb.py will exit if vdb is not found in the VE.

Updated Linux binaries
